move endpoints to their own package
